+++ packages/main/processes/andValue.js
@@ -1,4 +1,4 @@
-import map from '../signals/processes/map'
+import map from 'zen-signals/map'
 
 export default (otherValue) =>
   map((value) => ({value, ...otherValue}))