+++ packages/main/processes/hash.js
@@ -1,4 +1,4 @@
-import map from '../signals/processes/map'
+import map from 'zen-signals/map'
 import tail from 'ramda/src/tail'
 
 export default