+++ packages/main/processes/height.js
@@ -1,5 +1,5 @@
-import pipe from '../signals/pipe'
-import map from '../signals/processes/map'
+import pipe from 'ramda/src/pipe'
+import map from 'zen-signals/map'
 import target from './target'
 import prop from 'ramda/src/prop'