+++ packages/main/processes/position.js
@@ -6,7 +6,7 @@
  *
  */
 
-import map from '../signals/processes/map'
+import map from 'zen-signals/map'
 import props from 'ramda/src/props'
 
 export default map(props(['pageX', 'pageY']))