+++ packages/main/processes/preventDefault.js
@@ -9,7 +9,7 @@
  *
  */
 
-import map from '../signals/processes/map'
+import map from 'zen-signals/map'
 
 export default map((ev) => {
   ev.preventDefault()