+++ packages/main/processes/valueAndKeyCode.js
@@ -1,4 +1,4 @@
-import map from '../signals/processes/map'
+import map from 'zen-signals/map'
 
 export default
   map((ev) => [ev.target.value, ev.keyCode])