+++ packages/main/processes/valueOnEnter.js
@@ -1,4 +1,4 @@
-import filterMap from '../signals/processes/filterMap'
+import filterMap from 'zen-signals/filterMap'
 
 export default filterMap((ev) => ev.keyCode === 13 && ev.target.value.trim())