+++ packages/main/signals/processes/bind.js
@@ -1,5 +1,0 @@
-import build from './build'
-
-export default build((handler, next, value) => {
-  handler(value)(next)
-})